pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cda"
version = "0.1.0"
description = "CDA Project"
dependencies = [
    "fastapi",
    "uvicorn",
    "pyseto",
    "python-dotenv",
    "httpx",
    "aiosqlite",
]

[project.optional-dependencies]
test = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cda*"]